backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "risk-manager-pro"
version = "0.1.0"
description = "AI-driven operational risk management platform for private aviation safety"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "PyJWT>=2.9.0",
    "cryptography>=46.0.5",
    "httpx>=0.28.0",
    "aiohttp>=3.9.0",
    "azure-identity>=1.19.0",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.24.0",
    "openai>=1.57.0",
    "alembic>=1.14.0",
    "tenacity>=9.0.0",
    "python-multipart>=0.0.22",
    "pypdf>=6.7.5",
    "python-docx>=1.1.2",
    "tiktoken>=0.8.0",
    "slowapi>=0.1.9",
    "msal>=1.31.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "azure-ai-documentintelligence>=1.0.0",
    "pymupdf>=1.25.0",
    "azure-communication-email>=1.0.0",
    "jinja2>=3.1.4",
    "azure-monitor-opentelemetry>=1.6.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-asyncpg>=0.48b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
